# OpenAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "F1 API",
    "description": "An easier way to use the [ergast.com](https://ergast.com/mrd/) F1 API, with correct types!",
    "license": {
      "name": "MIT",
      "url": "https://github.com/niek-o/F1API/blob/main/LICENSE.md"
    },
    "version": "1.5.0",
    "x-logo": {
      "url": "https://upload.wikimedia.org/wikipedia/commons/f/f2/New_era_F1_logo.png"
    }
  },
  "paths": {
...
    "/results/race/{season}/{race}": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Race Results",
        "operationId": "get_race_results",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Race",
              "type": "integer"
            },
            "name": "race",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceResults"
                },
                "example": {
                  "results": [
                    {
                      "number": 1,
                      "position": 1,
                      "positionText": "1",
                      "points": 26,
                      "Driver": {
                        "driverId": "max_verstappen",
                        "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                        "givenName": "Max",
                        "familyName": "Verstappen",
                        "dateOfBirth": "1997-09-30",
                        "nationality": "Dutch",
                        "permanentNumber": 33,
                        "code": "VER"
                      },
                      "Constructor": {
                        "constructorId": "red_bull",
                        "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                        "name": "Red Bull",
                        "nationality": "Austrian"
                      },
                      "grid": 1,
                      "laps": 72,
                      "status": "Finished",
                      "Time": {
                        "time": "1:36:42.773",
                        "millis": "5802773"
                      },
                      "FastestLap": {
                        "rank": 1,
                        "lap": 62,
                        "Time": {
                          "time": "1:13.652"
                        },
                        "AverageSpeed": {
                          "units": "kph",
                          "speed": 208.173
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Race results not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/results/qualifying/{season}/{race}": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Qualifying Results",
        "operationId": "get_qualifying_results",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Race",
              "type": "integer"
            },
            "name": "race",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualifyingResults"
                },
                "example": {
                  "results": [
                    {
                      "number": 1,
                      "position": 1,
                      "Driver": {
                        "driverId": "max_verstappen",
                        "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                        "givenName": "Max",
                        "familyName": "Verstappen",
                        "dateOfBirth": "1997-09-30",
                        "nationality": "Dutch",
                        "permanentNumber": 33,
                        "code": "VER"
                      },
                      "Constructor": {
                        "constructorId": "red_bull",
                        "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                        "name": "Red Bull",
                        "nationality": "Austrian"
                      },
                      "Q1": "1:19.295",
                      "Q2": "1:18.793",
                      "Q3": "1:27.999"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Qualifying results not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/results/standings/drivers/{season}": {
      "get": {
        "tags": [
          "Results",
          "Season"
        ],
        "summary": "Get Driver Standings By Season",
        "operationId": "get_driver_standings_by_season",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStandings"
                },
                "example": {
                  "standings": [
                    {
                      "position": 1,
                      "positionText": "1",
                      "points": 335,
                      "wins": 11,
                      "Driver": {
                        "driverId": "max_verstappen",
                        "url": "https://://en.wikipedia.org/wiki/Max_Verstappen",
                        "givenName": "Max",
                        "familyName": "Verstappen",
                        "dateOfBirth": "1997-09-30",
                        "nationality": "Dutch",
                        "permanentNumber": 33,
                        "code": "VER"
                      },
                      "Constructors": [
                        {
                          "constructorId": "red_bull",
                          "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                          "name": "Red Bull",
                          "nationality": "Austrian"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Standings not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/results/standings/constructors/{season}": {
      "get": {
        "tags": [
          "Results",
          "Season"
        ],
        "summary": "Get Constructor Standings By Season",
        "operationId": "get_constructor_standings_by_season",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructorStandings"
                },
                "example": {
                  "standings": [
                    {
                      "position": 1,
                      "positionText": "1",
                      "points": 545,
                      "wins": 12,
                      "Constructor": {
                        "constructorId": "red_bull",
                        "url": "https://://en.wikipedia.org/wiki/Red_Bull_Racing",
                        "name": "Red Bull",
                        "nationality": "Austrian"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Standings not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AverageSpeed": {
        "title": "AverageSpeed",
        "required": [
          "units",
          "speed"
        ],
        "type": "object",
        "properties": {
          "units": {
            "title": "Units",
            "type": "string"
          },
          "speed": {
            "title": "Speed",
            "type": "number"
          }
        }
      },
...
      "Constructor": {
        "title": "Constructor",
        "required": [
          "constructorId",
          "url",
          "name",
          "nationality"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "constructorId": {
            "title": "Constructorid",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nationality": {
            "title": "Nationality",
            "type": "string"
          }
        }
      },
      "ConstructorStanding": {
        "title": "ConstructorStanding",
        "required": [
          "position",
          "positionText",
          "points",
          "wins",
          "Constructor"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "positionText": {
            "title": "Positiontext",
            "type": "string"
          },
          "points": {
            "title": "Points",
            "type": "integer"
          },
          "wins": {
            "title": "Wins",
            "type": "integer"
          },
          "Constructor": {
            "$ref": "#/components/schemas/Constructor"
          }
        }
      },
      "ConstructorStandings": {
        "title": "ConstructorStandings",
        "required": [
          "standings"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "standings": {
            "title": "Standings",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConstructorStanding"
            }
          }
        }
      },
...
      "Driver": {
        "title": "Driver",
        "required": [
          "driverId",
          "url",
          "givenName",
          "familyName",
          "dateOfBirth",
          "nationality"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "driverId": {
            "title": "Driverid",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "givenName": {
            "title": "Givenname",
            "type": "string"
          },
          "familyName": {
            "title": "Familyname",
            "type": "string"
          },
          "dateOfBirth": {
            "title": "Dateofbirth",
            "type": "string"
          },
          "nationality": {
            "title": "Nationality",
            "type": "string"
          },
          "permanentNumber": {
            "title": "Permanentnumber",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          }
        }
      },
      "DriverStanding": {
        "title": "DriverStanding",
        "required": [
          "position",
          "positionText",
          "points",
          "wins",
          "Driver",
          "Constructors"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "positionText": {
            "title": "Positiontext",
            "type": "string"
          },
          "points": {
            "title": "Points",
            "type": "integer"
          },
          "wins": {
            "title": "Wins",
            "type": "integer"
          },
          "Driver": {
            "$ref": "#/components/schemas/Driver"
          },
          "Constructors": {
            "title": "Constructors",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Constructor"
            }
          }
        }
      },
      "DriverStandings": {
        "title": "DriverStandings",
        "required": [
          "standings"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "standings": {
            "title": "Standings",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverStanding"
            }
          }
        }
      },
...
      "FastestLap": {
        "title": "FastestLap",
        "required": [
          "rank",
          "lap",
          "Time",
          "AverageSpeed"
        ],
        "type": "object",
        "properties": {
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "lap": {
            "title": "Lap",
            "type": "integer"
          },
          "Time": {
            "$ref": "#/components/schemas/Time"
          },
          "AverageSpeed": {
            "$ref": "#/components/schemas/AverageSpeed"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
...
      "QualifyingResult": {
        "title": "QualifyingResult",
        "required": [
          "number",
          "position",
          "Driver",
          "Constructor"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "Driver": {
            "$ref": "#/components/schemas/Driver"
          },
          "Constructor": {
            "$ref": "#/components/schemas/Constructor"
          },
          "Q1": {
            "title": "Q1",
            "type": "string"
          },
          "Q2": {
            "title": "Q2",
            "type": "string"
          },
          "Q3": {
            "title": "Q3",
            "type": "string"
          }
        }
      },
      "QualifyingResults": {
        "title": "QualifyingResults",
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "results": {
            "title": "Results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualifyingResult"
            }
          }
        }
      },
      "RaceResult": {
        "title": "RaceResult",
        "required": [
          "number",
          "position",
          "positionText",
          "points",
          "Driver",
          "Constructor",
          "grid",
          "laps",
          "status"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "positionText": {
            "title": "Positiontext",
            "type": "string"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "Driver": {
            "$ref": "#/components/schemas/Driver"
          },
          "Constructor": {
            "$ref": "#/components/schemas/Constructor"
          },
          "grid": {
            "title": "Grid",
            "type": "integer"
          },
          "laps": {
            "title": "Laps",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "Time": {
            "$ref": "#/components/schemas/Time"
          },
          "FastestLap": {
            "$ref": "#/components/schemas/FastestLap"
          }
        }
      },
      "RaceResults": {
        "title": "RaceResults",
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "results": {
            "title": "Results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RaceResult"
            }
          }
        }
      },
...
      "Time": {
        "title": "Time",
        "required": [
          "time"
        ],
        "type": "object",
        "properties": {
          "time": {
            "title": "Time",
            "type": "string"
          },
          "millis": {
            "title": "Millis",
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}